refactor(Web Page): move the board page header to a separate component
