+++ src/KanbanBoard.Frontend/src/pages/Board/BoardHeader/styles.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const Header = styled.header`
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    gap: 10px;
+    padding: 20px 0;
+`;
+
+export const BoardTitle = styled.h2`
+    margin-right: 50px;
+    margin: 0 50px 0 0;
+`;
